% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVICatMix.R
\name{runVICatMix}
\alias{runVICatMix}
\title{runVICatMix}
\usage{
runVICatMix(data, K, alpha, maxiter = 2000, tol = 5e-08)
}
\arguments{
\item{data}{- data frame/matrix with N rows of observations, and P columns of covariates}

\item{K}{- maximum number of clusters}

\item{alpha}{- Dirichlet prior parameter}

\item{maxiter}{- maximum number of interations}

\item{tol}{- convergence parameter}
}
\value{
A list.
}
\description{
runVICatMix
}
