% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVICatMixAvg.R
\name{runVICatMixAvg}
\alias{runVICatMixAvg}
\title{runVICatMixAvg}
\usage{
runVICatMixAvg(
  data,
  K,
  alpha,
  maxiter = 2000,
  tol = 5e-08,
  inits = 25,
  loss = "VoIcomp",
  parallel = FALSE,
  cores = getOption("mc.cores", 2L)
)
}
\arguments{
\item{data}{- data frame/matrix with N rows of observations, and P columns of covariates}

\item{K}{- maximum number of clusters}

\item{alpha}{- Dirichlet prior parameter}

\item{maxiter}{- maximum number of interations}

\item{tol}{- convergence parameter}

\item{inits}{- number of initialisations in the co-clustering matrix. Default 25}

\item{loss}{- loss function used in co-clustering matrix. Default VoIcomp. Options are "VoIavg", "VoIcomp" and "medv".}

\item{parallel}{- whether to use parallelisation to run initialisations}

\item{cores}{- Uses same default as mcclapply; user can specify number of cores for parallelisation if parallel = TRUE. Package automatically uses the user's parallel backend if one has already been registered.}
}
\value{
A list containing the averaged labels, as well as the results of each mixture model initialisation.
}
\description{
runVICatMixAvg
}
