% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVICatMixVarSel.R
\name{runVICatMixVarSel}
\alias{runVICatMixVarSel}
\title{runVICatMixVarSel}
\usage{
runVICatMixVarSel(data, K, alpha, a, maxiter = 2000, tol = 5e-08, outcome = NA)
}
\arguments{
\item{data}{- data frame/matrix with N rows of observations, and P columns of covariates}

\item{K}{- maximum number of clusters}

\item{alpha}{- Dirichlet prior parameter}

\item{a}{- hyperprior variable selection parameter}

\item{maxiter}{- maximum number of iterations}

\item{tol}{- convergence parameter}

\item{outcome}{- optional outcome variable. Default is NA; having an outcome allows for semi-supervised profile regression}
}
\value{
A list.
}
\description{
runVICatMixVarSel
}
