% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVICatMixVarSelAvg.R
\name{runVICatMixVarSelAvg}
\alias{runVICatMixVarSelAvg}
\title{runVICatMixVarSelAvg}
\usage{
runVICatMixVarSelAvg(
  data,
  K,
  alpha,
  a,
  maxiter = 2000,
  tol = 5e-08,
  inits = 25,
  loss = "VoIcomp",
  var_threshold = 0.95,
  parallel = FALSE,
  cores = getOption("mc.cores", 2L)
)
}
\arguments{
\item{data}{- data frame/matrix with N rows of observations, and P columns of covariates}

\item{K}{- maximum number of clusters}

\item{alpha}{- Dirichlet prior parameter}

\item{a}{- hyperprior variable selection parameter}

\item{maxiter}{- maximum number of interations}

\item{tol}{- convergence parameter}

\item{inits}{- number of initialisations in the co-clustering matrix. Default 25}

\item{loss}{- loss function used in co-clustering matrix. Default VoIcomp. Options are "VoIavg", "VoIcomp" and "medv".}

\item{var_threshold}{- treshold for selection rate for determining selected variables across the averaged model. Options are 0 < n <= 1 for a threshold. Default is 0.95.}

\item{parallel}{- whether to use parallelisation to run initialisations}

\item{cores}{- Uses same default as mcclapply; user can specify number of cores for parallelisation if parallel = TRUE. Package automatically uses the user's parallel backend if one has already been registered.}
}
\value{
A list containing the averaged labels, the averaged selected variables, as well as the results of each mixture model initialisation.
}
\description{
runVICatMixVarSelAvg
}
