\name{VICatMix-package}
\alias{VICatMix-package}
\alias{VICatMix}
\docType{package}
\title{
\packageTitle{VICatMix}
}
\description{
\packageDescription{VICatMix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{VICatMix}
\packageIndices{VICatMix}
~~ An overview of how to use the package, including important functions ~~
}
\author{
\packageAuthor{VICatMix}

Maintainer: \packageMaintainer{VICatMix}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
